refactor: restructure provider
