+++ src/AppProviders.tsx
@@ -14,7 +14,7 @@ import { Provider as ReduxProvider } from 'react-redux'
 import { HashRouter } from 'react-router-dom'
 import { PersistGate } from 'redux-persist/integration/react'
 import { ScrollToTop } from 'Routes/ScrollToTop'
-import { SwapperProvider } from 'components/Trade/swapperProvider'
+import { SwapperProvider } from 'components/Trade/SwapperProvider/swapperProvider'
 import { AppProvider } from 'context/AppProvider/AppContext'
 import { BrowserRouterProvider } from 'context/BrowserRouterProvider/BrowserRouterProvider'
 import { FoxEthProvider } from 'context/FoxEthProvider/FoxEthProvider'
