+++ src/components/Trade/SelectAccount.tsx
@@ -7,7 +7,8 @@ import { AssetAccountRow } from 'components/AssetAccounts/AssetAccountRow'
 import { Card } from 'components/Card/Card'
 import { SlideTransition } from 'components/SlideTransition'
 import { Text } from 'components/Text'
-import { SwapperActionType, useSwapperState } from 'components/Trade/swapperProvider'
+import { useSwapperState } from 'components/Trade/SwapperProvider/swapperProvider'
+import { SwapperActionType } from 'components/Trade/SwapperProvider/types'
 import { TradeRoutePaths } from 'components/Trade/types'
 import { WithBackButton } from 'components/Trade/WithBackButton'
 import { selectAccountIdsByAssetId, selectAssetById } from 'state/slices/selectors'
