+++ src/components/Trade/Trade.tsx
@@ -2,8 +2,9 @@ import type { AssetId } from '@shapeshiftoss/caip'
 import { useEffect, useState } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import { MemoryRouter, useLocation } from 'react-router-dom'
-import type { SwapperState } from 'components/Trade/swapperProvider'
-import { SwapperActionType, useSwapperState } from 'components/Trade/swapperProvider'
+import { useSwapperState } from 'components/Trade/SwapperProvider/swapperProvider'
+import type { SwapperState } from 'components/Trade/SwapperProvider/types'
+import { SwapperActionType } from 'components/Trade/SwapperProvider/types'
 
 import { useDefaultAssets } from './hooks/useDefaultAssets'
 import { TradeRoutes } from './TradeRoutes/TradeRoutes'
